requirements.txt
absl-py==0.9.0
attrs==19.3.0
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
cycler==0.10.0
decorator==4.4.2
Deprecated==1.2.10
future==0.18.2
google-auth==1.15.0
google-auth-oauthlib==0.4.1
grpcio==1.29.0
idna==2.9
imagecodecs==2020.2.18
imageio==2.8.0
importlib-metadata==1.6.0
joblib==0.15.1
kiwisolver==1.2.0
Markdown==3.2.2
matplotlib==3.2.1
more-itertools==8.3.0
networkx==2.4
numpy==1.19.0rc2
oauthlib==3.1.0
opencv-python==4.2.0.34
packaging==20.3
pandas==1.0.3
Pillow==7.1.2
pluggy==0.13.1
protobuf==3.12.0
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
pytest==5.4.2
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
-e git+https://github.com/lessw2020/Ranger-Deep-Learning-Optimizer@a170e3dac6443763961d0cb7a16294e4f45f93e4#egg=ranger
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-image==0.17.2
scikit-learn==0.23.1
scipy==1.4.1
SimpleITK==1.2.4
six==1.14.0
tensorboard==2.2.1
tensorboard-plugin-wit==1.6.0.post3
threadpoolctl==2.0.0
tifffile==2020.5.11
torch==1.6.0+cu101
torchsummary==1.5.1
torchvision==0.7.0+cu101
tornado==6.0.4
tqdm==4.46.0
urllib3==1.25.9
wcwidth==0.1.9
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
